Extracted AccumulateInlineVisitor
